+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## CHANGELOG
 
 ### 2.2.0 (25-Jul-2025)
-- 
+- Improved the GlobalProtect VPN IP detection logic
 
 ### 2.1.0 (24-Jul-2025)
 - Added an `operationMode` of "debug" to specifically enable swiftDialog debugging